// src/screens/fintechApp/payments/payout/interface.tsx
export interface KycFormData {
    firstName: string;
    lastName: string;
    email: string;
    phoneNumber: string;
    phoneCode: string;
    birthDate: string;
    occupation: string;
    country: string;
    address: string;
    docType: string;
    docNumber: string;
    expiryDate: string;
    frontId: string;
    backId: string;
}

export interface KycRequirements {
    kyc: {
        requirement: string;
    };
}
// ...
export interface AddressItem {
    id: string;
    favoriteName: string;
    name: string;
    [key: string]: any;
}

export interface PhoneCodeItem {
    code: string;
    name: string;
}

export interface FileNames {
    frontId: string | null;
    backId: string | null;
}

export interface ImagesLoader {
    frontId: boolean;
    backId: boolean;
}

export interface UserInfo {
    id: string;
    userName: string;
    [key: string]: any;
}

export interface ReduxState {
    userReducer: {
        userDetails: UserInfo;
    };
}

export interface RouteParams {
    VaultData?: {
        productId: string;
    };
}

export interface NavigationProp {
    navigate: (screen: string, params?: any) => void;
}

export interface PersonalKycFormProps {
    route: {
        params: RouteParams;
    };
}
// ...
export interface AddressObject {
    id: string;
    customerId: string;
    favoriteName: string;
    addressType: string;
    country: string;
    state: string;
    city: string;
    addressLine1: string;
    addressLine2: string;
    postalCode: string;
    phoneNumber: string;
    phoneCode: string;
    email: string;
    isDefault: boolean;
    createdBy: string;
    town: string;
    createdDate: Date;
}
// ...
export interface LookupData {
    Occupations?: Array<{ name: string; code: string }>;
    Country?: Array<{ name: string; code: string }>;
    documentTypes?: Array<{ name: string; code: string }>;
}

export interface AddressData {
    id: string;
    favoriteName: string;
    [key: string]: unknown;
}